Assistant checkpoint: Fix syntax error and run migration/upgrade
Assistant generated file changes:
- app/blueprints/products/product_inventory_routes.py: Fix syntax error in FIFO entries query

---

User prompt:

ok. migrate and upgrade

+++ app/blueprints/products/product_inventory_routes.py
@@ -271,7 +271,7 @@ def get_sku_fifo_status(sku_id):
             InventoryHistory.expiration_date.is_(None),  # Non-perishable
             InventoryHistory.expiration_date >= today    # Not expired yet
         )
-    ).order_by(InventoryHistory.timestamp.asc()).all()ll()
+    ).order_by(InventoryHistory.timestamp.asc()).all()
 
     # Get expired FIFO entries (frozen, with remaining quantity)
     expired_entries = InventoryHistory.query.filter(
